% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{mod_data}
\alias{mod_data}
\title{siteDrought_data server function}
\usage{
mod_data(input, output, session, main_data, main_date_range, lang)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{main_data}{main_data table}

\item{main_date_range}{absolute date range in main data}

\item{lang}{lang reactive}
}
\description{
siteDrought_data server function
}
