% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_helpOutput.R
\name{mod_helpOutput}
\alias{mod_helpOutput}
\title{mod_helpOutput and siteDrought_data}
\usage{
mod_helpOutput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny module to create and populate the data inputs
}
