% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_saveOutput.R
\name{mod_save}
\alias{mod_save}
\title{siteDrought_data server function}
\usage{
mod_save(
  input,
  output,
  session,
  data_reactives,
  main_data_reactives,
  main_data,
  lang
)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_reactives, main_data_reactives}{necessary reactives form other modules}

\item{main_data}{main_data table}

\item{lang}{lang reactive}
}
\description{
siteDrought_data server function
}
