% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_helpOutput.R
\name{mod_help}
\alias{mod_help}
\title{siteDrought_data server function}
\usage{
mod_help(input, output, session, main_data_reactives, lang)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{main_data_reactives}{reactives needed from other modules}

\item{lang}{lang reactive}
}
\description{
siteDrought_data server function
}
